FogBugz #492768: i2c: dw: temporary workaround for i2c clocks
This is a temporary workaround that:
* will not be submitted upstream
* will be reverted when the clock driver is fixed

Currently there are issues with writing to the stratix 10 clock
controller registers.  This commit allows the i2c controller to work
without doing any clock controller register access since we know the
clocks are enabled and we know what frequency they are.

This workaround is enabled by a setting in the stratix 10 dts, so it
will not affect other platforms that don't have this setting.

Signed-off-by: Alan Tull <atull@kernel.org>
---
v3: clarify explanation that this workaround is targeted to s10
    changed FB# to track the clock register issue
v2: add a setting in the device tree so this can be s10-only